requirements.txt
pandas==2.2.3
numpy==2.1.3
pyyaml==6.0.2